# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Anki Backend system. This API provides endpoints for managing decks, notes, cards, and study sessions.",
        "title": "Anki Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@anki.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/notes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "List notes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.NoteResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Create a note",
                "parameters": [
                    {
                        "description": "Note creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateNoteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.NoteResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Get note by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.NoteResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Update note",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateNoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.NoteResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Delete note",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/notes/{id}/tags": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Add tag to note",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tag request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AddTagRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/notes/{id}/tags/{tag}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "notes"
                ],
                "summary": "Remove tag from note",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tag name",
                        "name": "tag",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.AddTagRequest": {
            "type": "object",
            "required": [
                "tag"
            ],
            "properties": {
                "tag": {
                    "description": "Tag name to add",
                    "type": "string",
                    "example": "importante"
                }
            }
        },
...
        "request.CreateNoteRequest": {
            "type": "object",
            "required": [
                "deck_id",
                "fields_json",
                "note_type_id"
            ],
            "properties": {
                "deck_id": {
                    "description": "ID of the deck where cards will be created",
                    "type": "integer",
                    "example": 1
                },
                "fields_json": {
                    "description": "Fields data as a JSON string (e.g., '{\"Front\": \"Hello\", \"Back\": \"Olá\"}')",
                    "type": "string",
                    "example": "{\"Front\": \"cat\", \"Back\": \"gato\"}"
                },
                "note_type_id": {
                    "description": "ID of the note type (template)",
                    "type": "integer",
                    "example": 1
                },
                "tags": {
                    "description": "List of tags for the note",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"idiomas\"",
                        " \"animal\"]"
                    ]
                }
            }
        },
...
        "request.UpdateNoteRequest": {
            "type": "object",
            "required": [
                "fields_json"
            ],
            "properties": {
                "fields_json": {
                    "description": "Updated fields data as a JSON string",
                    "type": "string",
                    "example": "{\"Front\": \"cat\", \"Back\": \"gato (felino)\"}"
                },
                "tags": {
                    "description": "Updated list of tags",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"idiomas\"",
                        " \"mamífero\"]"
                    ]
                }
            }
        },
...
        "response.NoteResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp when created",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "fields_json": {
                    "description": "Fields data as a JSON string",
                    "type": "string",
                    "example": "{\"Front\": \"cat\", \"Back\": \"gato\"}"
                },
                "guid": {
                    "description": "GUID for synchronization",
                    "type": "string",
                    "example": "abc-123"
                },
                "id": {
                    "description": "Unique identifier for the note",
                    "type": "integer",
                    "example": 1
                },
                "note_type_id": {
                    "description": "ID of the note type",
                    "type": "integer",
                    "example": 1
                },
                "tags": {
                    "description": "List of tags",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"idiomas\"]"
                    ]
                },
                "updated_at": {
                    "description": "Timestamp when last updated",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}